Move get_command_path() to separate module
