+++ executorlib/shared/command.py
@@ -1,0 +1,14 @@
+import os
+
+
+def get_command_path(executable: str) -> str:
+    """
+    Get path of the backend executable script
+
+    Args:
+        executable (str): Name of the backend executable script, either mpiexec.py or serial.py
+
+    Returns:
+        str: absolute path to the executable script
+    """
+    return os.path.abspath(os.path.join(__file__, "..", "..", "backend", executable))
